+++ packages/serve/src/createConfig.ts
@@ -1,4 +1,8 @@
+import { utils } from 'webpack-cli';
+
 import { devServerOptionsType } from './types';
+
+const { logger } = utils;
 
 /**
  *
@@ -10,6 +14,16 @@ import { devServerOptionsType } from './types';
  */
 export default function createConfig(args): devServerOptionsType {
     const options = { ...args };
+    let isDevServer4 = false,
+        devServerVersion;
+    try {
+        // eslint-disable-next-line node/no-extraneous-require
+        devServerVersion = require('webpack-dev-server/package.json').version;
+    } catch (err) {
+        logger.error(`You need to install 'webpack-dev-server' for running 'webpack serve'.\n${err}`);
+        process.exit(2);
+    }
+    isDevServer4 = devServerVersion.startsWith('4');
 
     if (options.clientLogging) {
         options.client = {
@@ -18,8 +32,7 @@ export default function createConfig(args): devServerOptionsType {
         // clientLogging is not a valid devServer option
         delete options.clientLogging;
     }
-
-    if (options.hotOnly) {
+    if (isDevServer4 && options.hotOnly) {
         options.hot = 'only';
         // hotOnly is not a valid devServer option
         delete options.hotOnly;