+++ packages/serve/src/index.ts
@@ -1,6 +1,8 @@
-import WebpackCLI from 'webpack-cli';
+import WebpackCLI, { utils } from 'webpack-cli';
 import startDevServer from './startDevServer';
 import parseArgs from './parseArgs';
+
+const { logger } = utils;
 
 /**
  *
@@ -10,6 +12,13 @@ import parseArgs from './parseArgs';
  * @returns {Function} invokes the devServer API
  */
 export default function serve(...args: string[]): void {
+    try {
+        // eslint-disable-next-line node/no-extraneous-require
+        require('webpack-dev-server');
+    } catch (err) {
+        logger.error(`You need to install 'webpack-dev-server' for running 'webpack serve'.\n${err}`);
+        process.exit(2);
+    }
     const cli = new WebpackCLI();
 
     const { webpackArgs, devServerArgs } = parseArgs(cli, args);