+++ packages/webpack-cli/lib/utils/flag-defaults.js
@@ -1,23 +1,25 @@
-const cacheDefaults = (finalConfig, parsedArgs) => {
+const cacheDefaults = (finalConfig, parsedArgs, outputOptions) => {
     // eslint-disable-next-line no-prototype-builtins
     const hasCache = finalConfig.hasOwnProperty('cache');
     let cacheConfig = {};
-    if (hasCache && parsedArgs.config) {
+    if (hasCache && (parsedArgs.config || (outputOptions && outputOptions.defaultConfig))) {
         if (finalConfig.cache && finalConfig.cache.type === 'filesystem') {
             cacheConfig.buildDependencies = {
-                config: parsedArgs.config,
+                config: parsedArgs.config || [outputOptions.defaultConfig],
             };
+        } else {
+            cacheConfig = finalConfig.cache;
         }
         return { cache: cacheConfig };
     }
     return cacheConfig;
 };
 
-const assignFlagDefaults = (compilerConfig, parsedArgs) => {
+const assignFlagDefaults = (compilerConfig, parsedArgs, outputOptions) => {
     if (Array.isArray(compilerConfig)) {
-        return compilerConfig.map((config) => cacheDefaults(config, parsedArgs));
+        return compilerConfig.map((config) => cacheDefaults(config, parsedArgs, outputOptions));
     }
-    return cacheDefaults(compilerConfig, parsedArgs);
+    return cacheDefaults(compilerConfig, parsedArgs, outputOptions);
 };
 
 module.exports = assignFlagDefaults;