+++ test/bail/bail.test.js
@@ -4,15 +4,17 @@ const { run, runWatch } = require('../utils/test-utils');
 
 describe('bail and watch warning', () => {
     it('should not log warning in not watch mode', async () => {
-        const { stderr, stdout } = await run(__dirname, ['-c', 'bail-webpack.config.js']);
+        const { stderr, stdout, exitCode } = await run(__dirname, ['-c', 'bail-webpack.config.js']);
 
+        expect(exitCode).toEqual(0);
         expect(stderr).not.toContain(`You are using "bail" with "watch". "bail" will still exit webpack when the first error is found.`);
         expect(stdout).toBeTruthy();
     });
 
     it('should not log warning in not watch mode without the "bail" option', async () => {
-        const { stderr, stdout } = await run(__dirname, ['-c', 'no-bail-webpack.config.js']);
+        const { stderr, stdout, exitCode } = await run(__dirname, ['-c', 'no-bail-webpack.config.js']);
 
+        expect(exitCode).toEqual(0);
         expect(stderr).not.toContain(`You are using "bail" with "watch". "bail" will still exit webpack when the first error is found.`);
         expect(stdout).toBeTruthy();
     });