+++ test/config/defaults/multiple-location/multiple-location-config.test.js
@@ -5,8 +5,9 @@ const { run } = require('../../../utils/test-utils');
 
 describe('multiple dev config files with webpack.config.js', () => {
     it('Uses webpack.config.development.js', (done) => {
-        const { stdout, stderr } = run(__dirname, [], false);
+        const { stdout, stderr, exitCode } = run(__dirname, [], false);
         expect(stderr).toBeFalsy();
+        expect(exitCode).toEqual(0);
         expect(stdout).not.toBe(undefined);
         stat(resolve(__dirname, './binary/dev.folder.js'), (err, stats) => {
             expect(err).toBe(null);