+++ test/env/object/object-env.test.js
@@ -8,7 +8,12 @@ const { run } = require('../../utils/test-utils');
 
 describe('env object', () => {
     it('is able to set env for an object', () => {
-        run(__dirname);
+        const { stderr, stdout, exitCode } = run(__dirname);
+
+        expect(exitCode).toBe(0);
+        expect(stderr).toBeFalsy();
+        expect(stdout).toBeTruthy();
+
         const executable = path.join(__dirname, './bin/main.js');
         const bundledScript = spawnSync('node', [executable]);
         expect(bundledScript.stdout).toBe('environment is development');