+++ test/node/node.test.js
@@ -8,10 +8,12 @@ const { run } = require('../utils/test-utils');
 // throws different error from what we manually see
 describe('node flags', () => {
     it('is able to pass the options flags to node js', async (done) => {
-        const { stdout, stderr } = await run(__dirname, ['--output-path', './bin'], false, [
+        const { stdout, stderr, exitCode } = await run(__dirname, ['--output-path', './bin'], false, [
             `--require=${resolve(__dirname, 'bootstrap.js')}`,
             `--require=${resolve(__dirname, 'bootstrap2.js')}`,
         ]);
+
+        expect(exitCode).toBe(0);
         expect(stdout).toContain('---from bootstrap.js---');
         expect(stdout).toContain('---from bootstrap2.js---');
         expect(stderr).toBeFalsy();
@@ -23,18 +25,24 @@ describe('node flags', () => {
     });
 
     it('throws an error on supplying unknown flags', async () => {
-        const { stderr } = await run(__dirname, [], false, ['--unknown']);
+        const { stderr, exitCode } = await run(__dirname, [], false, ['--unknown']);
+
+        expect(exitCode).not.toBe(0);
         expect(stderr).toContain('bad option');
     });
 
     it('throws an error if no values were supplied with --max-old-space-size', async () => {
-        const { stderr, stdout } = await run(__dirname, [], false, ['--max-old-space-size']);
+        const { stderr, stdout, exitCode } = await run(__dirname, [], false, ['--max-old-space-size']);
+
+        expect(exitCode).not.toBe(0);
         expect(stderr).toContain('value for flag --max-old-space-size');
         expect(stdout).toBeFalsy();
     });
 
     it('throws an error if an illegal value was supplied with --max-old-space-size', async () => {
-        const { stderr, stdout } = await run(__dirname, [], true, ['--max_old_space_size=1024a']);
+        const { stderr, stdout, exitCode } = await run(__dirname, [], true, ['--max_old_space_size=1024a']);
+
+        expect(exitCode).not.toBe(0);
         expect(stderr).toContain('Error: illegal value for flag --max_old_space_size=1024a of type size_t');
         expect(stdout).toBeFalsy();
     });