+++ test/optimization/optimization.test.js
@@ -5,16 +5,18 @@ const { run } = require('../utils/test-utils');
 
 describe('optimization option in config', () => {
     it('should work with mangleExports disabled', () => {
-        const { stdout, stderr } = run(__dirname, [], false);
+        const { stdout, stderr, exitCode } = run(__dirname, [], false);
         // Should throw when webpack is less than 5
         if (!version.startsWith('5')) {
             expect(stderr).toContain("configuration.optimization has an unknown property 'mangleExports'");
+            expect(exitCode).toBe(2);
         } else {
             // Should apply the provided optimization to the compiler
             expect(stdout).toContain('mangleExports: false');
             // check that the output file exists
             expect(fs.existsSync(join(__dirname, '/dist/main.js'))).toBeTruthy();
             expect(stderr).toBeFalsy();
+            expect(exitCode).toBe(0);
         }
     });
 });