+++ test/stats/config/stats.test.js
@@ -6,7 +6,9 @@ const { version } = require('webpack');
 
 describe('stats flag with config', () => {
     it('should compile without stats flag', () => {
-        const { stderr, stdout } = run(__dirname, []);
+        const { stderr, stdout, exitCode } = run(__dirname, []);
+
+        expect(exitCode).toBe(0);
         expect(stderr).toBeFalsy();
         if (version.startsWith('5')) {
             expect(stdout).toContain(`stats: { preset: 'normal' }`);
@@ -15,7 +17,9 @@ describe('stats flag with config', () => {
         }
     });
     it('should compile with stats flag', () => {
-        const { stderr, stdout } = run(__dirname, ['--stats', 'errors-warnings']);
+        const { stderr, stdout, exitCode } = run(__dirname, ['--stats', 'errors-warnings']);
+
+        expect(exitCode).toBe(0);
         expect(stderr).toBeFalsy();
         if (version.startsWith('5')) {
             expect(stdout).toContain(`stats: { preset: 'errors-warnings' }`);