+++ test/zero-config/entry-absent/zero-config.test.js
@@ -2,9 +2,10 @@ const { run } = require('../../utils/test-utils');
 
 describe('Zero Config tests', () => {
     it('runs when config and entry are both absent', () => {
-        const { stdout, stderr } = run(__dirname, [], false);
+        const { stdout, stderr, exitCode } = run(__dirname, [], false);
         // Entry file is absent, should log the Error from the compiler
         expect(stdout).toContain("Error: Can't resolve './src'");
+        expect(exitCode).toBe(1);
         expect(stderr).toBeFalsy();
     });
 });